pageからプレゼンテーションの責務を分離
