+++ src/components/pages/article/index.ts
@@ -1,0 +1,1 @@
+export * from "./Article";
